Add support for async closing of files and dirs
Provides some initial work for #1, even though it still does not
support asynchronous drops.
