+++ CHANGELOG.md
@@ -3,7 +3,18 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added `TempDir` for automatically deleted temporary directories.
+- Added `close` methods to `TempFile` and `TempDir` allow for manual file system interactions.
+- Added `AsyncClose` trait for manual closing / deletion of `AsyncFile` and `AsyncDir` instances.
+  This is gated behind the `async-trait` crate feature and disabled by default.
+
 ## [0.5.0] - 2023-12-06
+
+[0.5.0]: https://github.com/sunsided/async-tempfile-rs/releases/tag/0.5.0
 
 ### Changed
 
@@ -16,6 +27,8 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Some unnecessary heap allocations were removed.
 
 ## [0.4.0] - 2023-06-16
+
+[0.4.0]: https://github.com/sunsided/async-tempfile-rs/releases/tag/0.4.0
 
 ### Added
 
@@ -44,6 +57,3 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Internal
 
 - 🎉 Initial release.
-
-[0.5.0]: https://github.com/sunsided/async-tempfile-rs/releases/tag/0.5.0
-[0.4.0]: https://github.com/sunsided/async-tempfile-rs/releases/tag/0.4.0
